try reading more bytes...?

+++ iiogyro.cpp
@@ -175,7 +175,7 @@ short Gyro::readRaw(const std::string& fileName, bool& status)
 		return 0;
 	}
 
-	int bytesRead = fread(buff,sizeof(short),1,fd);
+	int bytesRead = fread(buff,sizeof(int),1,fd);
 	if (sizeof(short) != bytesRead)
 	{
 		std::cerr<<"failed to read from "<<fileName<<" (read "<<bytesRead<<" instead of "<<sizeof(short)<<")"<<std::endl;
